+++ src/main/java/br/dev/andrestamatto/identityhub/infrastructure/messaging/config/MessagingConfiguration.java
@@ -17,10 +17,12 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.JavaMailSenderImpl;
+import org.springframework.retry.annotation.EnableRetry;
 
 import java.util.List;
 
 @Configuration
+@EnableRetry
 @EnableConfigurationProperties(NotificationProperties.class)
 public class MessagingConfiguration {